+++ components/Layout.tsx
@@ -1,4 +1,5 @@
 import Header from './Header'
+import LinkifyComponent from './linkify'
 import Navigation from './Navigation'
 
 export default function Layout({ children }: { children: React.ReactNode }) {